--- controllers/testController.js
@@ -1,6 +1,0 @@
-var socket = require('../core/socket');
-
-var indexTest = function(req, res, next) {
-	socket.test();
-}
-module.exports = indexTest;
