+++ package.json
@@ -5,17 +5,8 @@
   "scripts": {
     "start": "node app.js"
   },
-  "devDependencies": {
-    "bower": "~1.3.12"
-  },
   "dependencies": {
-    "body-parser": "~1.13.1",
-    "cookie-parser": "~1.3.5",
-    "debug": "~2.2.0",
-    "express": "~4.13.0",
-    "jade": "~1.11.0",
-    "morgan": "~1.6.1",
-    "serve-favicon": "~2.3.0",
-    "socket.io": "~1.3.6"
+    "express": "^4.10.2",
+    "socket.io": "^1.3.6"
   }
 }
