+++ public/game.html
@@ -1,0 +1,25 @@
+<!doctype html>
+<html>
+  <head>
+    <title>Index</title>
+  </head>
+  <body>
+  <script src="/socket.io/socket.io.js"></script>
+  <script>
+    var socket = io();
+
+    socket.emit('joinRoom', window.location.hash);
+
+    $('form').submit(function(){
+      socket.emit('chat', $('#m').val());
+      $('#m').val('');
+      return false;
+    });
+
+    socket.on('chat', function(msg){
+      $('#messages').append($('<li>').text(msg));
+    });
+
+  </script>
+  </body>
+</html>
