--- views/layout.jade
@@ -1,8 +1,0 @@
-doctype html
-html
-  head
-    title= title
-    link(rel='stylesheet', href='/stylesheets/style.css')
-    block scripts
-  body
-    block content
