--- views/screen.jade
@@ -1,9 +1,0 @@
-extends layout
-
-block scripts
-  script( src='/socket.io/socket.io.js' )
-  script( src='/js/app.js' )
-
-block content
-  h1=id
-  h2#incomming-message
